+++ sdks/python/apache_beam/io/gcp/bigquery_avro_tools.py
@@ -87,7 +87,7 @@ def get_record_schema_from_dict_table_schema(
 
 
 def table_field_to_avro_field(table_field, namespace):
-  # type: (Dict[Text, Any]) -> Dict[Text, Any]
+  # type: (Dict[Text, Any], str) -> Dict[Text, Any]
 
   """Convert a BigQuery field to an avro field.