+++ sdks/python/apache_beam/transforms/sql_test.py
@@ -30,17 +30,18 @@ from past.builtins import unicode
 
 import apache_beam as beam
 from apache_beam import coders
-from apache_beam.options.pipeline_options import DebugOptions
 from apache_beam.options.pipeline_options import StandardOptions
 from apache_beam.testing.test_pipeline import TestPipeline
 from apache_beam.testing.util import assert_that
 from apache_beam.testing.util import equal_to
 from apache_beam.transforms.sql import SqlTransform
-from apache_beam.utils import subprocess_server
 
 SimpleRow = typing.NamedTuple(
     "SimpleRow", [("int", int), ("str", unicode), ("flt", float)])
 coders.registry.register_coder(SimpleRow, coders.RowCoder)
+
+Enrich = typing.NamedTuple("Enrich", [("int", int), ("metadata", unicode)])
+coders.registry.register_coder(Enrich, coders.RowCoder)
 
 
 @attr('UsesSqlExpansionService')
@@ -66,20 +67,8 @@ class SqlTransformTest(unittest.TestCase):
         --tests apache_beam.transforms.sql_test \\
         --test-pipeline-options="--runner=FlinkRunner"
   """
-  @staticmethod
-  def make_test_pipeline():
-    path_to_jar = subprocess_server.JavaJarServer.path_to_beam_jar(
-        ":sdks:java:extensions:sql:expansion-service:shadowJar")
-    test_pipeline = TestPipeline()
-    # TODO(BEAM-9238): Remove this when it's no longer needed for artifact
-    # staging.
-    test_pipeline.get_pipeline_options().view_as(DebugOptions).experiments = [
-        'jar_packages=' + path_to_jar
-    ]
-    return test_pipeline
-
   def test_generate_data(self):
-    with self.make_test_pipeline() as p:
+    with TestPipeline() as p:
       out = p | SqlTransform(
           """SELECT
             CAST(1 AS INT) AS `int`,
@@ -88,14 +77,14 @@ class SqlTransformTest(unittest.TestCase):
       assert_that(out, equal_to([(1, "foo", 3.14)]))
 
   def test_project(self):
-    with self.make_test_pipeline() as p:
+    with TestPipeline() as p:
       out = (
           p | beam.Create([SimpleRow(1, "foo", 3.14)])
           | SqlTransform("SELECT `int`, `flt` FROM PCOLLECTION"))
       assert_that(out, equal_to([(1, 3.14)]))
 
   def test_filter(self):
-    with self.make_test_pipeline() as p:
+    with TestPipeline() as p:
       out = (
           p
           | beam.Create([SimpleRow(1, "foo", 3.14), SimpleRow(2, "bar", 1.414)])
@@ -103,7 +92,7 @@ class SqlTransformTest(unittest.TestCase):
       assert_that(out, equal_to([(2, "bar", 1.414)]))
 
   def test_agg(self):
-    with self.make_test_pipeline() as p:
+    with TestPipeline() as p:
       out = (
           p
           | beam.Create([
@@ -124,6 +113,30 @@ class SqlTransformTest(unittest.TestCase):
                 AVG(`flt`) AS `avg`
               FROM PCOLLECTION GROUP BY `str`"""))
       assert_that(out, equal_to([("foo", 3, 3, 2), ("bar", 4, 8, 1.414)]))
+
+  def test_tagged_join(self):
+    with TestPipeline() as p:
+      enrich = (
+          p | "Create enrich" >> beam.Create(
+              [Enrich(1, "a"), Enrich(2, "b"), Enrich(26, "z")]))
+      simple = (
+          p | "Create simple" >> beam.Create([
+              SimpleRow(1, "foo", 3.14),
+              SimpleRow(26, "bar", 1.11),
+              SimpleRow(1, "baz", 2.34)
+          ]))
+      out = ({
+          'simple': simple, 'enrich': enrich
+      }
+             | SqlTransform(
+                 """
+              SELECT
+                simple.`int` AS `int`,
+                enrich.metadata AS metadata
+              FROM simple
+              JOIN enrich
+              ON simple.`int` = enrich.`int`"""))
+      assert_that(out, equal_to([(1, "a"), (26, "z"), (1, "a")]))
 
 
 if __name__ == "__main__":